requirements.txt
Django>=4.2,<5.0
djangorestframework
drf_yasg
django-filter
bcrypt
drf-yasg
djangorestframework-simplejwt
PyJWT
pytz
django-cors-headers
